+++ src/level/RLSelectSends.hpp
@@ -13,6 +13,6 @@ protected:
     void onAllSends(CCObject* sender);
     void onThreePlusSends(CCObject* sender);
     void onLegendarySends(CCObject* sender);
-    void onMostSents(CCObject* sender);
-    void onLeastSents(CCObject* sender);
+    void onOldestSend(CCObject* sender);
+    void onQueueBtn(CCObject* sender);
 };