+++ src/popup/RLCreditsPopup.cpp
@@ -391,14 +391,14 @@ void RLCreditsPopup::onInfo(CCObject* sender) {
         "### <cr>Begging for Layout Mod to ArcticWoof or any of the Layout "
         "Admins will be ignored and lower your chances of becoming a mod.</c>\n"
         "If you have any questions about the application process or the role, feel free to ask in the <cl>Rated Layouts Discord Server</c>.\n"
-        "All promotion will be decided by <cf>ArcticWoof</c> and usually announced in the server."
         "\r\n\r\n---\r\n\r\n"
         "### Moderator Responsibilities\n"
         "- Moderators are expected to be <cg>active in the community</c> and help maintain the quality of the <cl>Rated Layouts</c>.\n"
         "- This includes <co>suggesting levels, rating levels</c> and <cy>providing feedback</c> to level creators.\n"
         "- Moderators may also be asked to help with <cg>managing the community</c>, such as moderating the leaderboard section or assisting with events.\n"
         "\r\n\r\n---\r\n\r\n"
-        "If you are <cg>interested in becoming a layout moderator</c>, make sure to join the <cl>Rated Layouts Discord Server</c> and apply in the <cl>application form</c>!",
+        "If you are <cg>interested in becoming a layout moderator</c>, make sure to join the <cl>Rated Layouts Discord Server</c> and apply in the <cl>application form</c>!\n"
+        "#### <cy>*Do note that ArcticWoof does not control over the selection process of moderators.*</cy>",
         "OK")
         ->show();
 }